producer/requirements.txt
statsbombpy==1.11.0
kafka-python-ng==2.2.2
requests==2.31.0
pandas==2.0.3
